tests/wpt/tests/css/css-inline/text-box-trim/parsing/text-box-invalid.html

<!DOCTYPE html>
<title>text-box rejects invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#propdef-text-box">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
// invalid keyword
test_invalid_value('text-box', 'em');
test_invalid_value('text-box', 'leading');
test_invalid_value('text-box', 'trim-both leading');

// invalid grammars
test_invalid_value('text-box', 'normal normal');
test_invalid_value('text-box', 'normal none');
test_invalid_value('text-box', 'auto normal');
test_invalid_value('text-box', 'auto auto');
test_invalid_value('text-box', 'none none');
test_invalid_value('text-box', 'normal text');
test_invalid_value('text-box', 'text normal');
test_invalid_value('text-box', 'text cap');
test_invalid_value('text-box', 'text ex');
test_invalid_value('text-box', 'alphabetic text');
test_invalid_value('text-box', 'alphabetic cap');
test_invalid_value('text-box', 'cap none alphabetic');
test_invalid_value('text-box', 'cap trim-both alphabetic');
test_invalid_value('text-box', 'text none text');
test_invalid_value('text-box', 'text trim-both text');
</script>
